daemon-cli: move multicast transport verbs into daemon-cli crate
Extract subscribe/unsubscribe/publish/unpublish into
doublezero-daemon-cli per RFC-20 (malbeclabs/infra#1493, PR 7 of the
extraction stack). Verbs stay nested under 'doublezero multicast'.
Deletes the binary's command/helpers.rs and the servicecontroller.rs
remnant; the crate copies survive.
