+++ CHANGELOG.md
@@ -32,6 +32,7 @@ All notable changes to this project will be documented in this file.
 - Onchain programs
   - Restrict granting the `FOUNDATION` permission flag: a plain `PERMISSION_ADMIN` holder can no longer grant `FOUNDATION` (a privilege escalation). Only a `foundation_allowlist` member or an existing `FOUNDATION` holder may grant it, enforced independently of `RequirePermissionAccounts` so foundation members are never locked out.
 - CLI
+  - Move the multicast transport verbs (`multicast subscribe`/`unsubscribe`/`publish`/`unpublish`) into the `doublezero-daemon-cli` crate per RFC-20. Each now takes `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; informational/result lines route through the shared writer (stdout, previously the stderr spinner). They stay nested under `doublezero multicast` (not hoisted as top-level daemon verbs); onchain `multicast group` CRUD is unchanged. The binary's `command/helpers.rs` (`resolve_client_ip`) and the `servicecontroller.rs` remnant are deleted — the crate copies survive. Flags, output content, and exit codes are unchanged.
   - Move the `connect` verb into the `doublezero-daemon-cli` crate per RFC-20. It now takes `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; informational/result lines route through the shared writer (stdout, previously the stderr spinner), spinners stay on stderr, pre-flight diagnostics route through `tracing`, and device selection uses the crate's latency utilities (from #3995). The binary's `dzd_latency.rs` and the orphaned `check_doublezero` pre-flight are deleted. Flags, output content, `--verbose`, exit codes, and version-check semantics are unchanged. (#4010)
   - Move the `disconnect` verb into the `doublezero-daemon-cli` crate per RFC-20. It now takes `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; informational/result lines route through the shared writer, spinners stay on stderr, and diagnostics route through `tracing`. Behavior (flags, output, `--verbose`, `--no-wait`, version-check) is unchanged. (#4008)
   - Move the `latency` and `routes` verbs, plus the device-selection/latency-polling utilities and `resolve_client_ip`, into the `doublezero-daemon-cli` crate per RFC-20. Both verbs now take `&CliContext` + generic `&D: DaemonClient` + `&L: LedgerClient` + `&mut W` writer; diagnostics route through `tracing` and output through the shared writer helper. Behavior (flags, output, `--json`, version-check) is unchanged. (#3995)
