+++ client/doublezero/Cargo.toml
@@ -14,7 +14,6 @@ repository.workspace = true
 [dependencies]
 anyhow.workspace = true
 backon.workspace = true
-chrono.workspace = true
 clap.workspace = true
 clap_complete.workspace = true
 console.workspace = true
@@ -22,12 +21,6 @@ eyre.workspace = true
 futures.workspace = true
 futures-util.workspace = true
 libc.workspace = true
-http.workspace = true
-http-body-util.workspace = true
-hyper.workspace = true
-hyper-util.workspace = true
-hyperlocal.workspace = true
-indicatif.workspace = true
 mockall.workspace = true
 serde.workspace = true
 serde_json.workspace = true
