--- client/doublezero/src/command/mod.rs
@@ -1,2 +1,0 @@
-pub mod helpers;
-pub mod multicast;
